Updates based on review
- Moved the "advanced" tokenization to the core folder leaving utils/text as a more simple way to process text.
- Added an error when the user updates the text if context doesn't contain the bookmark. This should also be enforced in the frontend.
